requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
python-multipart
jinja2
sqlalchemy
alembic
itsdangerous
bcrypt
twilio
python-dotenv
psycopg2-binary
pytest
httpx
